Enhancements of Cache
Created AbstractCache,

AbstractCache creates a lot of the methods so the implementations don't have to waste their time copy pasting.
Examples are the CustomCache#estimatedSize and the CustomCache#containsAll method.

This results in the IdCacheImpl and the NameCacheImpl only containing get, add, and remove methods

Furthermore, added better packaging in the cache package.
Now contains handler, policy and cachers sub-policies.